Improve README structure, wording of task list
Waypoint/introduction section at the top should stay.

Reordered the

Not sure yet at this point if the task list will remain in its current form right here. It's certainly useful for teachers & tutors to understand the workshop at a glance, but for students, sections will definitely be more useful.